requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.3
pickle-mixin==1.0.2
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.2
scikit-surprise==1.1.1
scipy==1.7.1
six==1.16.0
threadpoolctl==2.2.0
Werkzeug==2.0.1
